% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_db.R
\name{addModel.ca}
\alias{addModel.ca}
\title{Method to the forecast/learn models}
\usage{
addModel.ca(conn = NULL, model = NULL, label = NULL, description = NULL,
  predictors = NULL, params = NULL, metrics = NULL)
}
\description{
Serializes and add the models in the CA database for future use
}
\details{
@param conn - BAConnection object obtained from connect API
 @param model - model object which wants to be stored for future use
 @param label - label for the model
 @param description - short description about the model
 @param params - model specific parameters to set for future execution
 @param metrics - A dataFrame with metrics about the internal forecast/learn model that can be displayed to the user
}
