% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_db.R
\name{load.ca}
\alias{load.ca}
\title{Method to load the table data}
\usage{
load.ca(conn = NULL, columns = NULL)
}
\arguments{
\item{conn}{- BAConnection object obtained from connect API}

\item{columns}{- vector of selected column names}
}
\description{
User can pass the selected column names as the data frame returned by this
method will have only the selected columns
}
