% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_db.R
\name{update.ca}
\alias{update.ca}
\title{Method to update the data in the table}
\usage{
update.ca(conn = NULL, dataframe = NULL, model_name = NULL,
  model_label = NULL)
}
\description{
Update the table with the values provided in the data frame. Below is the
structure of the data frame for WhereClause
}
\details{
@param conn - BAConnection object obtained from connect API
 @param dataframe - DataFrame for where Clause
 @param model_name - Name of the model by running which the dataframe has been created(required in case of score/forecast)
 @param model_label - Label of the model by running which the dataframe has been created(required in case of score/forecast)
}
