% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics_db.R
\name{connect.ca}
\alias{connect.ca}
\title{Get datasource connection}
\usage{
connect.ca(url = NULL, token = NULL, apiKey = NULL, tunnelHost)
}
\arguments{
\item{token}{- Token for the datasource
obtained from the App}

\item{apiKey}{- ApiKey for the user/company}

\item{baseUrl}{- CA server URL}
}
\description{
This is a function to obtain connection to a
datasource from the envision server by providing the token
obtained from the envision App for the specific datasource
}
