% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carriots_analytics.R
\name{update.ca}
\alias{update.ca}
\title{Method to update the data in the table}
\usage{
\method{update}{ca}(conn = NULL, dataframe = NULL, colname = NULL,
  type = NULL)
}
\description{
Update the table with the values provided in the data frame. Below is the
structure of the data frame for WhereClause
}
\details{
@param conn - BAConnection object obtained from connect API
 @param dataframe - DataFrame for where Clause
 @param colname -  colname in dataframe which is to be added in to table
 @param type - Data type of column, supports specific types  available in the conn$dataTypes list
}
